+++ src/error.rs
@@ -4,6 +4,7 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use dusk_bytes::{BadLength, Error as DuskBytesError, InvalidChar};
 use dusk_poseidon::Error as PoseidonError;
 
 use core::fmt;
@@ -32,6 +33,12 @@ pub enum Error {
     InvalidCommitment,
     /// Invalid Nonce
     InvalidNonce,
+    /// Dusk-bytes InvalidData error
+    InvalidData,
+    /// Dusk-bytes BadLenght error
+    BadLenght(usize, usize),
+    /// Dusk-bytes InvalidChar error
+    InvalidChar(char, usize),
 }
 
 impl From<PoseidonError> for Error {
@@ -43,5 +50,32 @@ impl From<PoseidonError> for Error {
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "Phoenix-Core Error: {:?}", &self)
+    }
+}
+
+impl From<Error> for DuskBytesError {
+    fn from(err: Error) -> Self {
+        match err {
+            Error::InvalidData => DuskBytesError::InvalidData,
+            Error::BadLenght(found, expected) => {
+                DuskBytesError::BadLength { found, expected }
+            }
+            Error::InvalidChar(ch, index) => {
+                DuskBytesError::InvalidChar { ch, index }
+            }
+            _ => unreachable!(),
+        }
+    }
+}
+
+impl BadLength for Error {
+    fn bad_length(found: usize, expected: usize) -> Self {
+        Error::BadLenght(found, expected)
+    }
+}
+
+impl InvalidChar for Error {
+    fn invalid_char(ch: char, index: usize) -> Self {
+        Error::InvalidChar(ch, index)
     }
 }